move ErrorBoundary to its container
