+++ catalog/app/app.tsx
@@ -4,7 +4,6 @@
 // Import all the third party stuff
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import { useLocation } from 'react-router-dom'
 import { createBrowserHistory as createHistory } from 'history'
 import * as M from '@material-ui/core'
 
@@ -12,10 +11,8 @@ import * as M from '@material-ui/core'
 import 'sanitize.css'
 
 // Import root app
-import Error from 'components/Error'
 import { ExperimentsProvider } from 'components/Experiments'
 import * as Intercom from 'components/Intercom'
-import Layout from 'components/Layout'
 import Placeholder from 'components/Placeholder'
 import App from 'containers/App'
 import * as Auth from 'containers/Auth'
@@ -28,7 +25,6 @@ import * as APIConnector from 'utils/APIConnector'
 import { GraphQLProvider } from 'utils/GraphQL'
 import { BucketCacheProvider } from 'utils/BucketCache'
 import * as Config from 'utils/Config'
-import { createBoundary } from 'utils/ErrorBoundary'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import * as Cache from 'utils/ResourceCache'
 import * as Sentry from 'utils/Sentry'
@@ -50,42 +46,6 @@ fontLoader('Roboto', 'Roboto Mono').then(() => {
   // reload doc when we have all custom fonts
   document.body.classList.add('fontLoaded')
 })
-
-interface ErrorBoundaryPlaceholderProps {
-  error: Error
-  info: any
-  reset: () => void
-}
-
-function ErrorBoundaryPlaceholder({ error, info, reset }: ErrorBoundaryPlaceholderProps) {
-  const location = useLocation()
-  const errorShown = React.useRef(false)
-  React.useEffect(() => {
-    if (!errorShown.current) {
-      errorShown.current = true
-      return
-    }
-    errorShown.current = false
-    reset()
-  }, [location.pathname, reset])
-
-  const sentry = Sentry.use()
-  React.useEffect(() => {
-    sentry('captureException', error, info)
-  }, [error, info, sentry])
-
-  return (
-    <Layout bare>
-      <Error headline="Unexpected Error" detail="Something went wrong" />
-    </Layout>
-  )
-}
-
-const ErrorBoundary = createBoundary(
-  (_: unknown, { reset }: { reset: () => void }) =>
-    (error: $TSFixMe, info: $TSFixMe) =>
-      <ErrorBoundaryPlaceholder error={error} info={info} reset={reset} />,
-)
 
 const history = createHistory()
 const MOUNT_NODE = document.getElementById('app')
@@ -125,7 +85,7 @@ const render = () => {
       [React.Suspense, { fallback: <Placeholder /> }],
       [Sentry.Loader, { userSelector: sentryUserSelector }],
       GraphQLProvider,
-      ErrorBoundary,
+      Errors.ErrorBoundary,
       Notifications.Provider,
       [APIConnector.Provider, { fetch, middleware: [Auth.apiMiddleware] }],
       [Auth.Provider, { checkOn: LOCATION_CHANGE, storage }],
@@ -152,7 +112,7 @@ const render = () => {
       AWS.Athena.Provider,
       AWS.S3.Provider,
       Notifications.WithNotifications,
-      ErrorBoundary,
+      Errors.ErrorBoundary,
       BucketCacheProvider,
       App,
     ),
