+++ catalog/app/containers/Errors/index.ts
@@ -1,1 +1,2 @@
+export { default as ErrorBoundary } from './ErrorBoundary'
 export { default as FinalBoundary } from './FinalBoundary'
